feat: move main into cmd/gocq
